refactor: extract lib/schema-validation.mjs
